This code contains the GUI that: Allows the user to take an infinite amount of inputs and would store those inputs into a Dictionary.
Flaws that could not be quickly fixed:
I wanted to add the functionality where you could go back to a previous row and change the values if need be, but that proved to have a lot of issues. So once the (Add Input) button is pressed, the previous row's values are stored inside the dictionary.

In future, if we build on this program, I'll create an Edit Row button on each line to change the values, but yesh